Added main_test.py
remove code of partB from main_data.py
added main_test.py for partB
